TCP part 4 and splitting of packet.lua: - some final fixes to TCP - getXYZPacket: optional bool argument for all packet types (see TCP   part 3 commit) - moved utility functions for a specific header to respective files in   proto/ - only functions for pkt remain in packet.lua - made the new files more structured and readable
