+++ examples/l3-multi-destinations.lua
@@ -34,7 +34,7 @@ function loadSlave(port, queue, minA, numIPs)
 	--- parse and check ip addresses
 	-- min UDP packet size for IPv6 is 66 bytes
 	-- 4 bytes subtracted as the CRC gets appended by the NIC
-	local packetLen = 80 - 4
+	local packetLen = 66 - 4
 	local ipv4 = true
 	local minIP
 
@@ -69,16 +69,6 @@ function loadSlave(port, queue, minA, numIPs)
 		pkt:fill{ ethSrc="90:e2:ba:2c:cb:02", ethDst="90:e2:ba:35:b5:81", 
 				  ipSrc="192.168.1.1", 
 				  ip6Src="fd06::1",
-				  tcpSYN=1,
-				  --tcpPSH=1,
-				  --tcpURG=1,
-				  --tcpRST=1,
-				  --tcpFIN=1,
-				  --tcpFlags=0x0,
-				  --tcpSeqNumber=0x1447,
-				  --tcpAckNumber=0x1337,
-				  --tcpACK=1, 
-				  --tcpUrgentPointer=0xeffe,
 				  pktLength=packetLen }
 	end)
