--- lua/include/ethernet.lua
@@ -1,8 +1,0 @@
-local eth = {}
-
---- Ethernet constants
-eth.TYPE_IP = 0x0800
-eth.TYPE_ARP = 0x0806 -- not implemented
-eth.TYPE_IP6 = 0x86dd
-
-return eth
