--- lua/include/ip.lua
@@ -1,7 +1,0 @@
-local ip = {}
-
---- IPv4 constants
-ip.PROTO_TCP = 0x06
-ip.PROTO_UDP = 0x11
-
-return ip
