--- lua/include/ip6.lua
@@ -1,8 +1,0 @@
-local ip6 = {}
-
---- IP6 constants
--- TODO nextHeader uses same constants as IP protocol -> use the values from ip.lua?
-ip6.PROTO_TCP = 0x06
-ip6.PROTO_UDP = 0x11
- 
-return ip6
